allocation fix

+++ fppoly/fppoly.c
@@ -167,7 +167,7 @@ expr_t * create_sparse_expr(double *coeff, double cst, size_t *dim, size_t size)
 	if(size>0){
 		expr->inf_coeff = (double *)malloc(size*sizeof(double));
 		expr->sup_coeff = (double *)malloc(size*sizeof(double));
-		expr->dim = (size_t *)malloc(size*sizeof(double));
+		expr->dim = (size_t *)malloc(size*sizeof(size_t));
 	}
 	else{
 		expr->inf_coeff = NULL;
